+++ src/rank9.rs
@@ -69,9 +69,9 @@ pub struct Rank9 {
 }
 
 impl Access<bool> for Rank9 {
-    fn get(&self, n: int) -> bool {
-        let word = self.buffer[n as uint / 64];
-        (word >> (n as uint % 64)) & 1 == 1
+    fn get(&self, n: uint) -> bool {
+        let word = self.buffer[n / 64];
+        (word >> (n % 64)) & 1 == 1
     }
 }
